requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
pgvector==0.2.5
alembic==1.13.1

# AWS
boto3==1.34.34
botocore==1.34.34
python-jose[cryptography]==3.3.0  # Cognito JWT

# HTTP / environmental APIs
httpx==0.26.0
aiohttp==3.9.3

# Utilities
python-dotenv==1.0.1
structlog==24.1.0
tenacity==8.2.3
numpy==1.26.4

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
pytest-cov==4.1.0
httpx==0.26.0  # for TestClient
moto[dynamodb]==5.0.1  # mock AWS services
factory-boy==3.3.0
faker==22.5.1
